Add a vendored version of `libiconv` for macOS…
…to match the version that `superstring` expects.

This is copied directly from what Homebrew installs into `$(brew --prefix)/opt/libiconv`, except that the install name has been changed on `libiconv.2.dylib` to `libiconv.2.dylib`. (When it's installed from Homebrew, its macOS library “install name” matches its original path, so that's where macOS looks for it.

As part of the build, we rename the library with `install_name_tool` so that it can be resolved relative to the location of `superstring.node`.

This can probably be simplified, but first let's see if it'll work.